# docs/api-contract.openapi.yaml
openapi: 3.0.3
info:
  title: Jiwambe Onboarding — Platform API
  description: |
    Upstream API consumed by **jiwambe-onboarding-app** via its Next.js BFF.

    - Human-readable contract: [`api-contract.md`](./api-contract.md)
    - BFF routes mirror these paths under `/api/*` (see markdown BFF map)
    - Money fields are server-computed; clients display only
  version: 1.0.0
  contact:
    name: Jiwambe Platform
servers:
  - url: "{jiwambeApiBaseUrl}"
    description: Includes `/api/v1` — set via `JIWAMBE_API_BASE_URL`
    variables:
      jiwambeApiBaseUrl:
        default: https://api.example.jiwambe.com/api/v1
...
security:
  - bearerAuth: []
paths:
...
  /onboarding/auth/password/forgot:
    post:
      tags: [OfficerAuth]
      summary: Request password reset email
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Generic success (no enumeration)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /onboarding/auth/password/reset/password:
    post:
      tags: [OfficerAuth]
      summary: Set new password after reset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reset_session_id, password]
              properties:
                reset_session_id:
                  type: string
                password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "410":
          $ref: "#/components/responses/ResetExpired"
...
  /payments/validate:
    post:
      tags: [Payments]
      summary: Poll STK or validate M-Pesa receipt code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentValidateRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentValidateResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    RateLimited:
      description: Rate limited
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorBody"
              - type: object
                properties:
                  retry_after_seconds:
                    type: integer
...
    ResetExpired:
      description: Reset token expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
...
    ErrorBody:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    PaymentValidateRequest:
      type: object
      required: [applicationReferenceCode]
      properties:
        applicationReferenceCode:
          type: string
        checkoutId:
          type: string
        mpesaReceipt:
          type: string
    PaymentValidateResponse:
      type: object
      properties:
        status:
          type: string
          enum: [verified, pending, failed]
        mpesaReceipt:
          type: string
          nullable: true
        depositPayment:
          type: object
          nullable: true
        applicationVersion:
          type: integer